Fixed the delete method in the 'PhysicalFileSystem' class

+++ src/manager/v2/instances/PhysicalFileSystem.ts
@@ -150,9 +150,33 @@ export class PhysicalFileSystem extends FileSystem
         this.type(ctx.context, path, (e, type) => {
             if(e)
                 return callback(Errors.ResourceNotFound);
+            console.log(path.toString());
             
             if(type.isDirectory)
-                fs.rmdir(realPath, callback);
+            {
+                if(ctx.depth === 0)
+                    return fs.rmdir(realPath, callback);
+
+                this.readDir(ctx.context, path, (e, files) => {
+                    let nb = files.length + 1;
+                    const done = (e ?: Error) => {
+                        if(nb < 0)
+                            return;
+
+                        if(e)
+                        {
+                            nb = -1;
+                            return callback(e);
+                        }
+                        
+                        if(--nb === 0)
+                            fs.rmdir(realPath, callback);
+                    }
+
+                    files.forEach((file) => this.delete(ctx.context, path.getChildPath(file), ctx.depth === -1 ? -1 : ctx.depth - 1, (e) => done(e)));
+                    done();
+                })
+            }
             else
                 fs.unlink(realPath, callback);
         })
